refactor(model)!: Move custom JsonInclude filters to central place
This makes the filter implementation a bit easier to find. While at it
increase visibility to `internal` to enable re-use in an upcoming
change.
