+++ model/src/main/kotlin/config/RepositoryConfiguration.kt
@@ -21,6 +21,10 @@ package org.ossreviewtoolkit.model.config
 
 import com.fasterxml.jackson.annotation.JsonInclude
 
+import org.ossreviewtoolkit.model.utils.CurationsFilter
+import org.ossreviewtoolkit.model.utils.ExcludesFilter
+import org.ossreviewtoolkit.model.utils.LicenseChoiceFilter
+import org.ossreviewtoolkit.model.utils.ResolutionsFilter
 import org.ossreviewtoolkit.utils.ort.ORT_REPO_CONFIG_FILENAME
 
 /**
@@ -65,30 +69,3 @@ data class RepositoryConfiguration(
     @JsonInclude(value = JsonInclude.Include.CUSTOM, valueFilter = LicenseChoiceFilter::class)
     val licenseChoices: LicenseChoices = LicenseChoices()
 )
-
-@Suppress("EqualsOrHashCode", "EqualsWithHashCodeExist") // The class is not supposed to be used with hashing.
-private class ExcludesFilter {
-    override fun equals(other: Any?): Boolean =
-        if (other is Excludes) other.paths.isEmpty() && other.scopes.isEmpty() else false
-}
-
-@Suppress("EqualsOrHashCode", "EqualsWithHashCodeExist") // The class is not supposed to be used with hashing.
-private class ResolutionsFilter {
-    override fun equals(other: Any?): Boolean =
-        other is Resolutions &&
-                other.issues.isEmpty() &&
-                other.ruleViolations.isEmpty() &&
-                other.vulnerabilities.isEmpty()
-}
-
-@Suppress("EqualsOrHashCode", "EqualsWithHashCodeExist") // The class is not supposed to be used with hashing.
-private class CurationsFilter {
-    override fun equals(other: Any?): Boolean =
-        other is Curations && other.licenseFindings.isEmpty() && other.packages.isEmpty()
-}
-
-@Suppress("EqualsOrHashCode", "EqualsWithHashCodeExist") // The class is not supposed to be used with hashing.
-private class LicenseChoiceFilter {
-    override fun equals(other: Any?): Boolean =
-        other is LicenseChoices && other.isEmpty()
-}
